Training of Artificial Intelligence

Function for training an AI with several configurations and options provided by several packages (e.g. mlr3, text2vec).

ai_train(
  basic_text_rep,
  text_model,
  save_text_model = TRUE,
  learner,
  normalize = FALSE,
  n_performance_estimation = 30,
  ratio_performance_estimation = 0.75,
  category_name,
  filter = "jmim",
  filter_ratio = 1,
  verbose = FALSE,
  na.rm = TRUE
)

Arguments

basic_text_rep

Textual representation from which the AI should learn. The object should be generated with the function create_basic_representation.

text_model

The model of the texts that should be used for training AI. This object is must be generated with create_text_model.

save_text_model

TRUE if the text_model should be saved within the trained AI. FALSE if not desired.

learner

Learner used for modeling the AI. It must be created with mlr3.

normalize

TRUE if the independent and depended variables should be normalized before training AI. See details for more information.

n_performance_estimation

Number of iterations for estimating the performance of AI via bootstrap.

ratio_performance_estimation

Ratio for splitting the data into training data and test data.

category_name

Name of the target variable that should be predicted.

filter

Name of the filter that should be used.

filter_ratio

Amount of features that should be included in training the AI after applying a filter. For including all features set 1.0.

verbose

TRUE if information on the progress should be printed to the console. FALSE if not desired.

na.rm

TRUE if the analysis should use only complete cases. FALSE if not.

Value

Function returns an object of class List that contains the trained learner and all necessary components for predicting new data.

Details

method:Currently the function supports four methods. "DTM" forces that AI to learn based in the document-term-matrix. "LSA forces that an LSA is performed based on the initial text representation. AI uses the solution of the LSA. "LDA" forces that an LDA is performed based on the initial text representation. AI uses the solution of the LDA. "GLOVE" performs the estimation of global vectors based on the initial text representation. The result is used for learning by weighting the the global vectors with the word counts of the document-term-matrix.

normalize: If TRUE all data is normalized by applying the following formula to all columns of the data.frame: $$(0.95*(x - min(x)) / (max(x) - min(x)))+0.01$$