Creates a model of texts

Function for creating a textual model that provides the structure for preparing data for training AI and prediction new data.

create_text_model(
  basic_rep,
  method = "DTM",
  min_termfreq = 2,
  min_docfreq = 2,
  dim = 10,
  max_iter = 5000,
  verbose = TRUE
)

Arguments

basic_rep

Basic textual representation generated with the function create_basic_representation.

method

Method for building an advanced text representation as input for the AI. See details for more information.

min_termfreq

Int for determining the minimal appearance of a term in the whole data set.

min_docfreq

Int for determining the minimal number of documents that must contain a term. If the term is used in a lower number of documents it is excluded from the analysis.

dim

Int representing the number of modeled dimensions. It is only used in combination with methods "LDA","LSA", and ' "GLOVE".

max_iter

Maximum number of iterations allowed for "LDA", "LSA", and "GLOVE".

verbose

TRUE if information on the progress should be printed to the console. FALSE if not desired.

Value

Returns a list which contains the method used, the text model transformation_matrix for further analysis, and a matrix for matching new data into the structure of the generated text model(match_f_v_trim).

Details

Currently the function supports four methods.

"DTM" forces that AI to learn based in the document-term-matrix.

"LSA" frces that an LSA is performed based on the initial text representation. AI uses the solution of the LSA.

"LDA" forces that an LDA is performed based on the initial text representation. AI uses the solution of the LDA

"GLOVE" performs the estimation of global vectors based on the initial text representation. The result is used for learning by weighting the the global vectors with the word counts of the document-term-matrix.