Get pre configurated learners

Function for getting pre-configurated learners for convenience.

get_configurated_learner(
  learner_name,
  use_smote = TRUE,
  autotuning = FALSE,
  smote_K = 1,
  smote_k_max = 6,
  smote_dup = 1,
  smote_dup_max = 5,
  inner_sampling,
  cr_optim,
  max_n_tuning
)

Arguments

learner_name

Name of the learner to be used. See notes for more details.

use_smote

TRUE if smote should be applied. FALSE if not.

autotuning

TRUE if the learner should be auto tuned while learning. FALSE if not desired.

smote_K

Default value for applying smote.

smote_k_max

Maximum value for applying smote wihtin auto tuning.

smote_dup

Default value for applying smote.

smote_dup_max

Maximum value for applying smote wihtin auto tuning

inner_sampling

Kind of inner sampling for auto tuning.

cr_optim

Criterion for deciding about the best set of hyperparameters.

max_n_tuning

Maximum number of iterations for auto tuning.